1。react路由引入中间件概念。2.增加全线系统。3.改太多忘了。

+++ app/Http/Controllers/Api/FileController.php
@@ -12,44 +12,64 @@ class FileController extends Controller
     /* CommonFunc 封装有一系列通用功能函数 */
     use CommonFunc;
 
-    function recieveFile(Request $request){
-        $files=$request->file();
-        $path=null;
-        $user=User::find($request->get('user'));
-        foreach ($files as $key=>$file){
-            $newName=md5(date('Ymd').$file->getClientOriginalName()).'.'.($file->getClientOriginalExtension());
-            $path = $file->storeAs('/user/'.$user->id,$newName , 'uploads');
-            $resource=new Resources();
-            $resource->name=$newName;
-            $resource->original_name=$file->getClientOriginalName();
-            $resource->size='not sure';
-            $resource->type=$file->getClientMimeType();
-            $resource->path=env('APP_URL').'/react/'.$path;
-           /* $resource->save();*/
-            $data['link']=$resource->path;
-            if($resource->save()){
-                $user->resources()->attach($resource->id);
-                return response()->json(['data'=>$data],200);
-            }else{
-               return response()->json(['error'=>'something wrong'],200);
+    function recieveFile(Request $request)
+    {
+        $files = $request->file();
+        $path = null;
+        $user=$request->user();
+        if($request->user()->tokenCan('files-upload')){
+            foreach ($files as $key => $file) {
+                $newName = md5(date('Ymd') . $file->getClientOriginalName()) . '.' . ($file->getClientOriginalExtension());
+                $path = $file->storeAs('/user/' . $user->id, $newName, 'uploads');
+                $resource = new Resources();
+                $resource->name = $newName;
+                $resource->original_name = $file->getClientOriginalName();
+                $resource->size = 'not sure';
+                $resource->type = $file->getClientMimeType();
+                $resource->path = env('APP_URL') . '/react/' . $path;
+                /* $resource->save();*/
+                $data['link'] = $resource->path;
+                if ($resource->save()) {
+                    $user->resources()->attach($resource->id);
+                    return response()->json(['data' => $data], 200);
+                } else {
+                    return response()->json(['error' => 'something wrong'], 200);
+                }
             }
+        }else{
+            $msg=['error'=>"You dont have permisiion"];
+            return response()->json($msg,403);
         }
+
     }
-    function userFiles(Request $request){
-        return $request->user()->resources;
+
+    function userImgs(Request $request){
+        $files=$request->user()->resources()->where('type','!=','video/mp4')->orderBy('id', 'desc')->paginate($request->get('pageSize'));
+        return $files;
+    }
+    function userVideos(Request $request){
+        $files=$request->user()->resources()->where('type','=','video/mp4')->orderBy('id', 'desc')->paginate($request->get('pageSize'));
+        return $files;
     }
     function delete(Request $request){
         $file=Resources::find($request->get('id'));
         $id=$file->id;$name=$file->name;
-        if($file->delete()){
-            Storage::disk('uploads')->delete($name);
-            $request->user()->resources()->detach($id);
-            $msg=['status'=>'success'];
-            return $msg;
+        $user=$request->user();
+        if($user->tokenCan('files-delete')){
+            if($file->delete()){
+                Storage::disk('uploads')->delete($name);
+                $request->user()->resources()->detach($id);
+                $msg=['success'=>'Delete success!'];
+                return $msg;
+            }else{
+                $msg=['error'=>'Delete Error!Something Wrong!'];
+                return $msg;
+            }
         }else{
-            $msg=['status'=>'false'];
+            $msg = ['error' => 'Delete Error!You Dont Have Permission!'];
             return $msg;
         }
+
 
     }
